+++ packages/insights-charts/source/js/charts/ColumnChart.js
@@ -224,6 +224,12 @@ class ColumnChart extends Chart {
 
     svg.selectAll(CSS.getClassSelector('series')).raise();
   }
+
+  destroy() {
+    if (this.tooltip) {
+      this.tooltip.destroy();
+    }
+  }
 }
 
 export default ColumnChart;