+++ packages/insights-charts/source/js/charts/CombinationChart.js
@@ -491,6 +491,12 @@ class CombinationChart extends Chart {
 
     svg.selectAll(CSS.getClassSelector('series')).raise();
   }
+
+  destroy() {
+    if (this.tooltip) {
+      this.tooltip.destroy();
+    }
+  }
 }
 
 export default CombinationChart;