+++ packages/insights-charts/source/js/charts/LineChart.js
@@ -214,6 +214,12 @@ class LineChart extends Chart {
 
     svg.selectAll(CSS.getClassSelector('series')).raise();
   }
+
+  destroy() {
+    if (this.tooltip) {
+      this.tooltip.destroy();
+    }
+  }
 }
 
 export default LineChart;