+++ packages/insights-charts/source/js/charts/ScatterChart.js
@@ -182,6 +182,12 @@ class ScatterChart extends Chart {
 
     svg.selectAll(CSS.getClassSelector('series')).raise();
   }
+
+  destroy() {
+    if (this.tooltip) {
+      this.tooltip.destroy();
+    }
+  }
 }
 
 export default ScatterChart;